Moved migration relevant functions to an own file
